pyproject.toml
[tool.poetry]
name = "grouping-bot"
version = "0.1.0"
description = ""
authors = ["Yumax-panda"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
pytest = "^7.4.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-isort = "^6.1.1"
flake8-bugbear = "^24.1.17"
flake8-builtins = "^2.2.0"
flake8-eradicate = "^1.5.0"
flake8-unused-arguments = "^0.0.13"
flake8-pytest-style = "^1.7.2"
pep8-naming = "^0.13.3"
mypy = "^1.8.0"
black = "^23.12.1"
isort = "^5.13.2"
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
test = "pytest -s -vv --cov=. --cov-branch --cov-report=html"
fmt = "task fmt-black && task fmt-isort"
fmt-black = "black grouping_bot tests"
fmt-isort = "isort grouping_bot tests"
lint = "task lint-black && task lint-flake8 && task lint-mypy"
lint-flake8 = "flake8 grouping_bot tests"
lint-mypy = "mypy grouping_bot tests"
lint-black = "black --check grouping_bot tests"
run-bot = "python grouping_bot/bot.py"
